+++ packages/docs/src/components/menus/LearnMenu.vue
@@ -26,8 +26,8 @@
             to: {
               name: 'Documentation',
               params: {
-                category: 'beginners-guide',
-                page: 'guide',
+                category: 'introduction',
+                page: 'what-is-vuetify',
               },
             },
           },